Implement get_navigation on GEVERClient
To be able to get the navigation of a GEVER, I introduce a new method called
"get_navigation", which lets you get the navigation of a given position in
GEVER but also on the root.

To ensure we have the same defaults `raw=False` as on every other method which
the apiclient exposes (done by the `@autowrap` decorator), I implemented the
method in the same manner.